+++ catalog/app/app.tsx
@@ -39,6 +39,7 @@ import * as APIConnector from 'utils/APIConnector'
 import * as GraphQL from 'utils/GraphQL'
 import { BucketCacheProvider } from 'utils/BucketCache'
 import GlobalAPI from 'utils/GlobalAPI'
+import log from 'utils/Logging'
 import * as NamedRoutes from 'utils/NamedRoutes'
 import * as Cache from 'utils/ResourceCache'
 import * as Store from 'utils/Store'
@@ -58,10 +59,15 @@ globalApi.attach(window)
 const GlobalAPIProvider = globalApi.getProvider()
 
 // listen for Roboto fonts
-fontLoader('Roboto', 'Roboto Mono').then(() => {
-  // reload doc when we have all custom fonts
-  document.body.classList.add('fontLoaded')
-})
+fontLoader('Roboto', 'Roboto Mono')
+  .then(() => {
+    // reload doc when we have all custom fonts
+    document.body.classList.add('fontLoaded')
+  })
+  .catch((error) => {
+    log.log('Failed to load fonts')
+    log.error(error)
+  })
 
 const MOUNT_NODE = document.getElementById('app')